+++ src/components/BreadcrumbMenu/BreadcrumbMenu.tsx
@@ -1,8 +1,15 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Outlet, Link, NavLink, useParams, matchPath } from 'react-router-dom';
-import { HiMenuAlt3 } from 'react-icons/hi';
+import { NavLink, useParams, matchPath } from 'react-router-dom';
 import useBreadcrumbs from 'use-react-router-breadcrumbs';
+import {
+  navLinkStyles,
+  activeStyle,
+  BreadcrumbArea,
+  CohortContainer,
+  CohortDash,
+  BreadcrumbContainer,
+  MenuIcon,
+} from './BreadcrumbMenu.styled';
 
 const DynamicBreadcrumb = () => {
   const { id } = useParams();
@@ -19,15 +26,6 @@ const excludePaths = ['/', '/admin/cohorts/:id'];
 
 const isRouteExcluded = (pathname: string) => {
   return excludePaths.find((path) => matchPath({ path }, pathname));
-};
-
-const navLinkStyles = {
-  color: 'white',
-};
-
-const activeStyle = {
-  fontWeight: 'bold',
-  fontSize: '1.125em',
 };
 
 const Breadcrumbs = () => {
@@ -60,46 +58,6 @@ const Breadcrumbs = () => {
     </>
   );
 };
-
-const BreadcrumbArea = styled.div`
-  background: ${({ theme }) => theme.background.primary};
-  margin: 1.5em;
-  height: 90px;
-  border-radius: 1em;
-  width: 90%;
-  display: flex;
-  flex-direction: column;
-  color: ${({ theme }) => theme.background.white};
-`;
-
-const CohortContainer = styled.section`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-const CohortDash = styled.section`
-  color: ${({ theme }) => theme.background.white};
-  margin: 1em;
-  font-size: 0.85em;
-  font-weight: 600;
-`;
-
-const BreadcrumbContainer = styled.span`
-  padding-left: 1.5em;
-  font-size: 0.75em;
-  color: ${({ theme }) => theme.background.white};
-`;
-
-const iconStyles = {
-  fontSize: '1.5rem',
-  margin: '.175em',
-  color: 'white',
-};
-
-export const MenuIcon = styled(HiMenuAlt3)`
-  ${iconStyles}
-`;
 
 export const BreadcrumbMenu = () => {
   return (
